Move selection related tests.
Part of reorganizing the tests requires grouping tests. One such group
is anything related to selections, e.g. column, row or point selections,
hyperslabs, etc.

This commit only moves code from one file to the next without any
changes.
